// src/api/generated/model/workoutSessionDetailResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */
import type { WorkoutSimpleResponse } from './workoutSimpleResponse';
import type { WorkoutSessionDetailResponseStatus } from './workoutSessionDetailResponseStatus';
import type { WorkoutExerciseSessionResponse } from './workoutExerciseSessionResponse';

export interface WorkoutSessionDetailResponse {
  id?: number;
  workout?: WorkoutSimpleResponse;
  weekWorkoutId?: number;
  status?: WorkoutSessionDetailResponseStatus;
  startedAt?: string;
  finishedAt?: string;
  workoutExerciseSessions?: WorkoutExerciseSessionResponse[];
}