// src/api/generated/model/workoutExerciseSetResultInputRequestWorkoutExerciseSetType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Fitness App API
 * Comprehensive API for fitness application with support for workout plans, exercises, and progress tracking. API provides OAuth2 authentication via Keycloak or JWT tokens.
 */

export type WorkoutExerciseSetResultInputRequestWorkoutExerciseSetType = typeof WorkoutExerciseSetResultInputRequestWorkoutExerciseSetType[keyof typeof WorkoutExerciseSetResultInputRequestWorkoutExerciseSetType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WorkoutExerciseSetResultInputRequestWorkoutExerciseSetType = {
  WARMUP: 'WARMUP',
  TOP_SET: 'TOP_SET',
  BACKOFF_SET: 'BACKOFF_SET',
  STRAIGHT_SET: 'STRAIGHT_SET',
  DROP_SET: 'DROP_SET',
} as const;